Update automations limitation
